+++ detectors/aws/ec2/ec2.go
@@ -25,7 +25,7 @@ import (
 
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/sdk/resource"
-	semconv "go.opentelemetry.io/otel/semconv/v1.12.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
 )
 
 type config struct {
@@ -104,12 +104,12 @@ func (detector *resourceDetector) Detect(ctx context.Context) (*resource.Resourc
 	attributes := []attribute.KeyValue{
 		semconv.CloudProviderAWS,
 		semconv.CloudPlatformAWSEC2,
-		semconv.CloudRegionKey.String(doc.Region),
-		semconv.CloudAvailabilityZoneKey.String(doc.AvailabilityZone),
-		semconv.CloudAccountIDKey.String(doc.AccountID),
-		semconv.HostIDKey.String(doc.InstanceID),
-		semconv.HostImageIDKey.String(doc.ImageID),
-		semconv.HostTypeKey.String(doc.InstanceType),
+		semconv.CloudRegion(doc.Region),
+		semconv.CloudAvailabilityZone(doc.AvailabilityZone),
+		semconv.CloudAccountID(doc.AccountID),
+		semconv.HostID(doc.InstanceID),
+		semconv.HostImageID(doc.ImageID),
+		semconv.HostType(doc.InstanceType),
 	}
 
 	m := &metadata{client: client}