+++ instrumentation/github.com/aws/aws-sdk-go-v2/otelaws/test/aws_test.go
@@ -44,7 +44,7 @@ func TestAppendMiddlewares(t *testing.T) {
 		expectedStatusCode int
 	}{
 		"invalidChangeBatchError": {
-			responseStatus: 500,
+			responseStatus: http.StatusInternalServerError,
 			responseBody: []byte(`<?xml version="1.0" encoding="UTF-8"?>
 		<InvalidChangeBatch xmlns="https://route53.amazonaws.com/doc/2013-04-01/">
 		  <Messages>
@@ -55,11 +55,11 @@ func TestAppendMiddlewares(t *testing.T) {
 			expectedRegion:     "us-east-1",
 			expectedError:      codes.Error,
 			expectedRequestID:  "b25f48e8-84fd-11e6-80d9-574e0c4664cb",
-			expectedStatusCode: 500,
+			expectedStatusCode: http.StatusInternalServerError,
 		},
 
 		"standardRestXMLError": {
-			responseStatus: 404,
+			responseStatus: http.StatusNotFound,
 			responseBody: []byte(`<?xml version="1.0"?>
 		<ErrorResponse xmlns="http://route53.amazonaws.com/doc/2016-09-07/">
 		  <Error>
@@ -73,11 +73,11 @@ func TestAppendMiddlewares(t *testing.T) {
 			expectedRegion:     "us-west-1",
 			expectedError:      codes.Error,
 			expectedRequestID:  "1234567890A",
-			expectedStatusCode: 404,
+			expectedStatusCode: http.StatusNotFound,
 		},
 
 		"Success response": {
-			responseStatus: 200,
+			responseStatus: http.StatusOK,
 			responseBody: []byte(`<?xml version="1.0" encoding="UTF-8"?>
 		<ChangeResourceRecordSetsResponse>
    			<ChangeInfo>
@@ -86,12 +86,12 @@ func TestAppendMiddlewares(t *testing.T) {
    		</ChangeInfo>
 		</ChangeResourceRecordSetsResponse>`),
 			expectedRegion:     "us-west-2",
-			expectedStatusCode: 200,
+			expectedStatusCode: http.StatusOK,
 		},
 	}
 
 	for name, c := range cases {
-		server := httptest.NewServer(http.HandlerFunc(
+		srv := httptest.NewServer(http.HandlerFunc(
 			func(w http.ResponseWriter, r *http.Request) {
 				w.WriteHeader(c.responseStatus)
 				_, err := w.Write(c.responseBody)
@@ -109,7 +109,7 @@ func TestAppendMiddlewares(t *testing.T) {
 				EndpointResolverWithOptions: aws.EndpointResolverWithOptionsFunc(
 					func(service, region string, _ ...interface{}) (aws.Endpoint, error) {
 						return aws.Endpoint{
-							URL:         server.URL,
+							URL:         srv.URL,
 							SigningName: "route53",
 						}, nil
 					},
@@ -151,6 +151,6 @@ func TestAppendMiddlewares(t *testing.T) {
 			assert.Contains(t, attrs, attribute.String("aws.operation", "ChangeResourceRecordSets"))
 		})
 
-		server.Close()
+		srv.Close()
 	}
 }