+++ instrumentation/github.com/gocql/gocql/otelgocql/example/client.go
@@ -48,7 +48,6 @@ import (
 	"go.opentelemetry.io/otel/metric/global"
 	"go.opentelemetry.io/otel/sdk/metric"
 	"go.opentelemetry.io/otel/sdk/metric/aggregation"
-	"go.opentelemetry.io/otel/sdk/metric/view"
 	"go.opentelemetry.io/otel/sdk/trace"
 
 	"go.opentelemetry.io/contrib/instrumentation/github.com/gocql/gocql/otelgocql"
@@ -126,76 +125,40 @@ func main() {
 	wg.Wait()
 }
 
-func views() ([]view.View, error) {
-	var vs []view.View
-	// TODO: Remove renames when the Prometheus exporter natively supports
-	// metric instrument name sanitation
-	// (https://github.com/open-telemetry/opentelemetry-go/issues/3183).
-	v, err := view.New(
-		view.MatchInstrumentName("db.cassandra.queries"),
-		view.WithRename("db_cassandra_queries"),
-	)
-	if err != nil {
-		return nil, err
+func views() []metric.View {
+	return []metric.View{
+		metric.NewView(
+			metric.Instrument{
+				Name: "db.cassandra.rows",
+			},
+			metric.Stream{
+				Aggregation: aggregation.ExplicitBucketHistogram{
+					Boundaries: []float64{0.001, 0.01, 0.1, 0.5, 1, 2, 5, 10},
+				},
+			},
+		),
+		metric.NewView(
+			metric.Instrument{Name: "db.cassandra.latency"},
+			metric.Stream{
+				Aggregation: aggregation.ExplicitBucketHistogram{
+					Boundaries: []float64{0.001, 0.01, 0.1, 0.5, 1, 2, 5, 10},
+				},
+			},
+		),
 	}
-	vs = append(vs, v)
-
-	v, err = view.New(
-		view.MatchInstrumentName("db.cassandra.rows"),
-		view.WithRename("db_cassandra_rows"),
-		view.WithSetAggregation(aggregation.ExplicitBucketHistogram{
-			Boundaries: []float64{0.001, 0.01, 0.1, 0.5, 1, 2, 5, 10},
-		}),
-	)
-	if err != nil {
-		return nil, err
-	}
-	vs = append(vs, v)
-
-	v, err = view.New(
-		view.MatchInstrumentName("db.cassandra.batch.queries"),
-		view.WithRename("db_cassandra_batch_queries"),
-	)
-	if err != nil {
-		return nil, err
-	}
-	vs = append(vs, v)
-
-	v, err = view.New(
-		view.MatchInstrumentName("db.cassandra.connections"),
-		view.WithRename("db_cassandra_connections"),
-	)
-	if err != nil {
-		return nil, err
-	}
-	vs = append(vs, v)
-
-	v, err = view.New(
-		view.MatchInstrumentName("db.cassandra.latency"),
-		view.WithRename("db_cassandra_latency"),
-		view.WithSetAggregation(aggregation.ExplicitBucketHistogram{
-			Boundaries: []float64{0.001, 0.01, 0.1, 0.5, 1, 2, 5, 10},
-		}),
-	)
-	if err != nil {
-		return nil, err
-	}
-	vs = append(vs, v)
-
-	return vs, nil
 }
 
 func initMetrics() error {
-	vs, err := views()
-	if err != nil {
-		return err
-	}
+	vs := views()
 
 	exporter, err := otelprom.New()
 	if err != nil {
 		return err
 	}
-	provider := metric.NewMeterProvider(metric.WithReader(exporter, vs...))
+	provider := metric.NewMeterProvider(
+		metric.WithReader(exporter),
+		metric.WithView(vs...),
+	)
 	global.SetMeterProvider(provider)
 
 	http.Handle("/", promhttp.Handler())